% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_loss.R
\name{change_loss}
\alias{change_loss}
\title{Changes the loss mark of a question}
\usage{
change_loss(x, a)
}
\arguments{
\item{x}{A question}

\item{a}{The new loss mark}
}
\value{
A question with the same parameters but with new loss mark
}
\description{
Changes the loss mark of a question
}
\examples{
question_1 <- gen_mc(2)
question_1 <- change_loss(question_1, 1)
question_2 <- change_loss(question_1, 0)
}
