% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compile questions into a test}
\usage{
compile(...)
}
\arguments{
\item{...}{A group of either questions or tests}
}
\value{
A test comprimising of the given questions, or a new test comprimising of a combination of tests.
}
\description{
Compile questions into a test
}
\examples{
question_1 <- gen_mc(2)
question_2 <- gen_match(2)
test_1 <- compile(question_1, question_2)
test_2 <- compile(question_1, question_1)
test_3 <- compile(test_1, test_2)
}
