% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_match.R
\name{gen_match}
\alias{gen_match}
\title{Creates a matching style question}
\usage{
gen_match(p, m = 1, l = 0, forall = TRUE)
}
\arguments{
\item{p}{Number of matches to be made for the question}

\item{m}{The score to be assigned if correct}

\item{l}{The score to be deducted if wrong}

\item{forall}{The logical condition describing whether the score is assigned if all the matches are correct, or if it is assigned for each correct match}
}
\value{
An object containing all the parameters of the question
}
\description{
Creates a matching style question
}
\examples{
question_1 <- gen_match(p = 5, m = 3)
question_2 <- gen_match(p = 6, l = 1, forall = FALSE)
}
