% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_mc.R
\name{gen_mc}
\alias{gen_mc}
\title{Creates a multiple choice style question}
\usage{
gen_mc(p, c = 1, m = 1, l = 0, forall = TRUE)
}
\arguments{
\item{p}{The number of possible answers for the question}

\item{c}{The number of correct answers for the question}

\item{m}{The score to be assigned if correct}

\item{l}{The score to be deducted if wrong}

\item{forall}{The logical condition describing whether the score is assigned if all the correct answers are chosen, or if it is assigned for each correct answer chosen}
}
\value{
An object containing all the parameters of the question
}
\description{
Creates a multiple choice style question
}
\examples{
question_1 <- gen_mc(p = 4, m = 3)
question_2 <- gen_mc(p = 6, c = 2, l = 1, forall = FALSE)
}
