% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_loss.R
\name{mark_loss}
\alias{mark_loss}
\title{Find the optimum loss mark}
\usage{
mark_loss(x)
}
\arguments{
\item{x}{A question or a test}
}
\value{
For a question, `mark_loss` will output the mark to be deducted
if answered incorrectly such that the expected mark from random guessing
is negative. For a test, it outputs the average loss mark.
}
\description{
Find the optimum loss mark
}
\examples{
question_1 <- gen_mc(2)
test_1 <- compile(question_1, question_1)
mark_loss(question_1)
mark_loss(test_1)
}
