% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_mark.R
\name{max_mark}
\alias{max_mark}
\title{Find the maximum mark}
\usage{
max_mark(x)
}
\arguments{
\item{x}{A test}
}
\value{
The maximum possible mark for that test.
}
\description{
Find the maximum mark
}
\examples{
question_1 <- gen_mc(2)
test_1 <- compile(question_1)
max_mark(test_1)
}
