% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{qsim}
\alias{qsim}
\title{Simulate a random guess at a question}
\usage{
qsim(x)
}
\arguments{
\item{x}{A question}
}
\value{
A score produced by simulating a random guess at a question under its defined parameters.
}
\description{
Simulate a random guess at a question
}
\examples{
question_1 <- gen_mc(2)
question_2 <- gen_match(4)
qsim(question_1)
qsim(question_2)
}
