% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_loss.R
\name{set_loss}
\alias{set_loss}
\title{Change the loss mark for every question in a test}
\usage{
set_loss(x, a)
}
\arguments{
\item{x}{A test}

\item{a}{The new loss mark}
}
\value{
An identical test but with every question having the new loss mark parameter.
}
\description{
Change the loss mark for every question in a test
}
\examples{
question_1 <- gen_mc(2)
test_1 <- compile(question_1)
test_1 <- set_loss(test_1, 0.5)
test_2 <- set_loss(test_1, 0)
}
