% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dist.R
\name{sim_dist}
\alias{sim_dist}
\title{Simulate the distribution of random guess scores for a test}
\usage{
sim_dist(x, n = 10000)
}
\arguments{
\item{x}{A test}

\item{n}{The number of simulations to be made. Increasing n will increase the accuracy of the simulation.}
}
\value{
An object containing the n results of the simulation and the maximum possible mark for the test. All values are rounded accordingly.
}
\description{
Simulate the distribution of random guess scores for a test
}
\examples{
question_1 <- gen_mc(2)
test_1 <- rep(question_1, 5)
distribution_1 <- sim_dist(test_1)
distribution_2 <- sim_dist(test_1, n = 500)
}
