% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dist.R
\name{summary.dist}
\alias{summary.dist}
\title{A summary of results from a distribution}
\usage{
\method{summary}{dist}(object, pass = 0.4, probs = TRUE, ...)
}
\arguments{
\item{object}{A simulated distribution}

\item{pass}{The percentage required in order to pass. Must be between 0 and 1.}

\item{probs}{A logic statement determining if the simulated probabilities should be printed.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Displays a variety of statistics from the simulated distribution including mean, median,
highest simulated mark, highest possible mark, the estimated probabilities of attaining each score,
the probability of passing based on estimated probabilities and a graphical representation of the
simulated distribution.
}
\description{
A summary of results from a distribution
}
\examples{
question_1 <- gen_mc(2)
test_1 <- rep(question_1, 5)
distribution_1 <- sim_dist(test_1)
summary(distribution_1)
summary(distribution_1, 0.6)
}
